Move session interface modification logic to the security manager for easy debugging and customization
